+++ R/check_prior.R
@@ -4,7 +4,8 @@
 #'
 #' @inheritParams effective_sample
 #' @param method Can be "gelman" or "lakeland". For the "gelman" method, if the SD of the posterior is more than 0.1 times the SD of the prior, then the prior is considered as informative. For the "lakeland" method, the prior is considered as informative if the posterior falls within the 95\% HDI of the prior.
-#' @param simulate_priors Should prior distributions be simulated using \code{simulate_prior} (default; faster) or sampled (slower, more accurate).
+#' @param simulate_priors Should prior distributions be simulated using
+#'   \code{\link{simulate_prior}} (default; faster) or sampled via \code{\link{unupdate}} (slower, more accurate).
 #' @examples
 #' \dontrun{
 #' library(bayestestR)
@@ -57,7 +58,7 @@ check_prior.brmsfit <- function(model, method = "gelman", simulate_priors = TRUE
         parameters = parameters
       )
   } else {
-    priors <- .update_to_priors(model, verbose = FALSE)
+    priors <- unupdate(model, verbose = FALSE)
     priors <-
       insight::get_parameters(
         priors,