+++ R/rope.R
@@ -185,18 +185,24 @@ rope.data.frame <- function(x, range = "default", ci = .89, ci_method = "HDI", v
 #' @rdname rope
 #' @export
 rope.emmGrid <- function(x, range = "default", ci = .89, ci_method = "HDI", verbose = TRUE, ...) {
-  xdf <- .clean_emmeans_draws(x)
+  xdf <- insight::get_parameters(x)
 
   dat <- rope(xdf, range = range, ci = ci, ci_method = ci_method, verbose = verbose, ...)
   attr(dat, "object_name") <- .safe_deparse(substitute(x))
   dat
 }
 
+#' @export
+rope.emm_list <- rope.emmGrid
+
 
 
 #' @rdname rope
 #' @export
 rope.BFBayesFactor <- function(x, range = "default", ci = .89, ci_method = "HDI", verbose = TRUE, ...) {
+  if (all(range == "default")) {
+    range <- rope_range(x)
+  }
   out <- rope(insight::get_parameters(x), range = range, ci = ci, ci_method = ci_method, verbose = verbose, ...)
   attr(out, "object_name") <- .safe_deparse(substitute(x))
   out
@@ -297,7 +303,7 @@ rope.stanreg <- function(x, range = "default", ci = .89, ci_method = "HDI", effe
     ...
   )
 
-  out <- .prepare_output(rope_data, insight::clean_parameters(x))
+  out <- .prepare_output(rope_data, insight::clean_parameters(x), inherits(x, "stanmvreg"))
 
   attr(out, "HDI_area") <- attr(rope_data, "HDI_area")
   attr(out, "object_name") <- .safe_deparse(substitute(x))