+++ R/update.bayesfactor_models.R
@@ -1,6 +1,6 @@
-#' Update bayesfactor_models
+#' Methods for \code{bayesfactor_models}
 #'
-#' @param object A \code{\link{bayesfactor_models}} object.
+#' @param object,x A \code{\link{bayesfactor_models}} object.
 #' @param subset Vector of model indices to keep or remove.
 #' @param reference Index of model to rereference to, or \code{"top"} to reference to the best model, or \code{"bottom"} to reference to the worst model.
 #' @param ... Currently not used.
@@ -19,6 +19,7 @@
 #' m
 #'
 #' update(m, reference = "bottom")
+#' as.matrix(m)
 #' }
 #' @export
 update.bayesfactor_models <- function(object, subset = NULL, reference = NULL, ...) {
@@ -35,6 +36,9 @@ update.bayesfactor_models <- function(object, subset = NULL, reference = NULL, .
   denominator <- attr(object, "denominator")
 
   if (!is.null(subset)) {
+    if (all(subset < 0)) {
+      subset <- seq_len(nrow(object))[subset]
+    }
     object_subset <- object[subset, ]
 
     if (denominator %in% subset) {
@@ -47,3 +51,19 @@ update.bayesfactor_models <- function(object, subset = NULL, reference = NULL, .
   }
   object
 }
+
+
+#' @rdname update.bayesfactor_models
+#' @export
+as.matrix.bayesfactor_models <- function(x, ...) {
+  x$BF <- log(x$BF)
+  out <- -outer(x$BF, x$BF, FUN = "-")
+  rownames(out) <- colnames(out) <- x$Model
+
+  out <- exp(out)
+
+  class(out) <- c("bayesfactor_models_matrix", class(out))
+  out
+}
+
+