+++ tests/testthat/test-rope.R
@@ -29,8 +29,8 @@ if (require("rstanarm", quietly = TRUE) && require("brms", quietly = TRUE)) {
   })
 
 
-  .runThisTest <- Sys.getenv("RunAllinsightTests") == "yes"
-  if (.runThisTest || Sys.getenv("USER") == "travis") {
+  .runThisTest <- Sys.getenv("RunAllbayestestRTests") == "yes"
+  if (.runThisTest) {
     if (require("insight")) {
       m <- insight::download_model("stanreg_merMod_5")
       p <- insight::get_parameters(m, effects = "all")
@@ -56,4 +56,12 @@ if (require("rstanarm", quietly = TRUE) && require("brms", quietly = TRUE)) {
       })
     }
   }
+}
+
+
+if (require("BayesFactor", quietly = TRUE)) {
+  mods <- regressionBF(mpg ~ am + cyl, mtcars, progress = FALSE)
+  rx <- suppressMessages(rope(mods))
+  expect_equal(rx$ROPE_high, -rx$ROPE_low)
+  expect_equal(rx$ROPE_high[1], 0.6026948)
 }