+++ NEWS.md
@@ -1,4 +1,23 @@
-# bayestestR 0.8.0
+# bayestestR 0.7.5
+
+## New functions
+
+* Added `as.matrix()` function for `bayesfactor_model` arrays.
+* `unupdate()`, a utility function to get Bayesian models un-fitted from the data, representing the priors only.
+
+## Changes to functions
+
+* `ci()` supports `emmeans` - both Bayesian and frequentist ( #312 - cross fix with `parameters`)
+
+## Bug fixes
+
+* Fixed issue with *default* rope range for `BayesFactor` models.
+* Fixed issue in collinearity-check for `rope()` for models with less than two parameters.
+* Fixed issue in print-method for `mediation()` with `stanmvreg`-models, which displays the wrong name for the response-value.
+* Fixed issue in `effective_sample()` for models with only one parameter.
+* `rope_range()` for `BayesFactor` models returns non-`NA` values ( #343 )
+
+# bayestestR 0.7.2
 
 ## New functions